% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\alias{uninstall}
\alias{validate}
\title{Install and uninstall 'rcli' to the R startup process}
\usage{
install(
  path = "~",
  backup = TRUE,
  overwrite = FALSE,
  quiet = FALSE,
  validate = TRUE
)

uninstall(path = "~", backup = TRUE, quiet = FALSE)

validate()
}
\arguments{
\item{path}{The path where to create / update the \file{.Rprofile} file.}

\item{backup}{If \code{TRUE}, a timestamped backup copy of the original file is
created before modifying / overwriting it, otherwise not.  If the backup
fails, then an error is produced and the R startup file is unmodified.}

\item{overwrite}{If the R startup file already exist, then \code{FALSE} (default)
appends the rcli code to the end of the file. is overwritten.  If \code{TRUE},
any pre-existing R startup file is overwritten.}

\item{quiet}{If \code{FALSE} (default), detailed messages are generated,
otherwise not.}

\item{validate}{If \code{TRUE} (default), the installation is validated.}
}
\value{
(invisible) The pathname of the R startup file modified.
}
\description{
Install and uninstall 'rcli' by appending / removing one line of code to
the \file{~/.Rprofile} file.
}
\section{Functions}{
\itemize{
\item \code{install}: injects\if{html}{\out{<div class="r">}}\preformatted{if (nzchar(Sys.getenv("R_CMD")) && require("rcli", quietly=TRUE)) rcli::r_cmd_call()
}\if{html}{\out{</div>}}

to the \file{.Rprofile} file (created if missing).

\item \code{uninstall}: Remove calls to \code{startup::startup()} and similar.
}}

