% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{is_debug_on}
\alias{is_debug_on}
\title{Checks whether debug is on or not}
\usage{
is_debug_on()
}
\value{
Returns \code{TRUE} is debug is enabled and \code{FALSE} otherwise.
}
\description{
Checks whether debug is on or not
}
\details{
The debug mode is when one of the \pkg{rcliaddons} is called, either
explicitly via argument \code{debug} or via environment variable
\code{R_RCLIADDONS_DEBUG}.
}
\keyword{internal}
