% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_cmd_call.R
\name{r_cmd_call}
\alias{r_cmd_call}
\title{\emph{(EXPERIMENTAL)} Performs a tweaked 'R CMD' call}
\usage{
r_cmd_call(
  extras = c("debug", "as", "config", "renviron"),
  args = commandArgs(trailingOnly = TRUE),
  unload = TRUE,
  debug = NA,
  envir = parent.frame(),
  dryrun = FALSE
)
}
\arguments{
\item{extras}{(character vector) Additional features that should be injected
to \verb{R CMD}.}

\item{args}{(optional) Command-line arguments after the \code{--args} option}

\item{unload}{If \code{TRUE}, then the package is unloaded afterward, otherwise
not.}

\item{debug}{If TRUE, debug information is outputted, otherwise not.}

\item{envir}{The environment where to evaluate any \R expression used
internally by the \verb{R CMD} call.}

\item{dryrun}{If TRUE, nothing is evaluated.}
}
\value{
(logical; invisible) TRUE or FALSE.
}
\description{
\emph{(EXPERIMENTAL)} Performs a tweaked 'R CMD' call
}
\section{Installation}{

This function should be called at the end of the \file{Rprofile} startup
file.  You can use \code{rcli::install()} to automatically append:\if{html}{\out{<div class="r">}}\preformatted{if (nzchar(Sys.getenv("R_CMD")) && require("rcli", quietly=TRUE)) rcli::r_cmd_call()
}\if{html}{\out{</div>}}

to your \file{~/.Rprofile} file.
}

\section{The \code{renviron} feature}{

When \code{extras} include \code{"renviron"}, this function will add support for the
\verb{--renviron=<name>} command-line option to \verb{R CMD build} and \verb{R CMD check}.
When specified, \verb{R CMD build --renviron=<name>} will run with
\code{R_BUILD_ENVIRON} set to \file{~/.R/<name>-build.Renviron}.
Similarly, \verb{R CMD check --renviron=<name>} will run with
\code{R_CHECK_ENVIRON} set to \file{~/.R/<name>-check.Renviron}.
}

\keyword{internal}
