\docType{package}
\name{RCLI}
\alias{RCLI}
\alias{RCLI-package}
\title{RCLI}
\description{
  A simple command line interface for R scripts, allowing
  arguments to be passed as named arguments to
  \code{Rscript} invocations from the command line, as well
  as a simple method for defining usage of an R script from
  the command line.
}

